requirements.txt
asgiref==3.4.1
Django==3.2.5
django-ratings==0.3.7
djangorestframework==3.12.4
Pillow==8.3.0
psycopg2-binary==2.9.1
pytz==2021.1
sqlparse==0.4.1
django-utils-six=2.0